refactor connection/monitoring
